refactor: move the modules related to file processing to the lib directory
