+++ lib/file/pdf-processing/core/index.js
