+++ lib/file/pdf-processing/strategy/default.js
