+++ lib/file/pdf-processing/strategy/index.js
