+++ lib/file/pdf-processing/strategy/vision.js
