+++ lib/file/split-mardown/core/splitter.js
