+++ lib/file/text-splitter.js
@@ -2,7 +2,7 @@
 
 import fs from 'fs';
 import path from 'path';
-import { getProjectRoot, ensureDir } from './db/base';
+import { getProjectRoot, ensureDir } from '../db/base';
 import { getProject } from '@/lib/db/projects';
 import { getChunkByProjectId, saveChunks } from '@/lib/db/chunks';
