+++ lib/util/domain-tree.js
@@ -2,15 +2,15 @@
  * 领域树处理模块
  * 用于处理领域树的生成、修订和管理
  */
-const LLMClient = require('./llm/core/index');
-const getLabelPrompt = require('./llm/prompts/label');
-const getLabelEnPrompt = require('./llm/prompts/labelEn');
-const getLabelRevisePrompt = require('./llm/prompts/labelRevise');
-const getLabelReviseEnPrompt = require('./llm/prompts/labelReviseEn');
-const { getProjectTocs } = require('./text-splitter');
-const { getTags, batchSaveTags } = require('./db/tags');
-const { extractJsonFromLLMOutput } = require('./llm/common/util');
-const { filterDomainTree } = require('./util/file');
+const LLMClient = require('../llm/core/index');
+const getLabelPrompt = require('../llm/prompts/label');
+const getLabelEnPrompt = require('../llm/prompts/labelEn');
+const getLabelRevisePrompt = require('../llm/prompts/labelRevise');
+const getLabelReviseEnPrompt = require('../llm/prompts/labelReviseEn');
+const { getProjectTocs } = require('../file/text-splitter');
+const { getTags, batchSaveTags } = require('../db/tags');
+const { extractJsonFromLLMOutput } = require('../llm/common/util');
+const { filterDomainTree } = require('./file');
 
 /**
  * 处理领域树生成或更新
